third_party/WebKit/LayoutTests/http/tests/websocket/invalid-encode-length-expected.txt

CONSOLE ERROR: line 21: WebSocket connection to 'ws://127.0.0.1:8880/invalid-encode-length?case=126_125' failed: Invalid frame header
CONSOLE ERROR: line 21: WebSocket connection to 'ws://127.0.0.1:8880/invalid-encode-length?case=126_0' failed: Invalid frame header
CONSOLE ERROR: line 21: WebSocket connection to 'ws://127.0.0.1:8880/invalid-encode-length?case=127_65535' failed: Invalid frame header
CONSOLE ERROR: line 21: WebSocket connection to 'ws://127.0.0.1:8880/invalid-encode-length?case=127_0' failed: Invalid frame header
Test whether WebSocket aborts the connection when it receives an unexpected encoding length frame.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Case 126_125: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case 126_0: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case 127_65535: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
Case 127_0: Test started.
onopen() was called.
PASS onerror() was called
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
